Attempt to Fix some broken links and images in ITALIAN.md
Hi, @Syknapse , I notice that some links and images are broken in the italian translation and also in the portuguese one.

So, I tried for fixing the images to add just a simple slash "/" before the image link and regarding the broken links, I simply remove the space between square and round brackets. As I'm concerned the reason for this issue may be that the translated files are stored in a different directory so the path for accessing the images is different.  I am new here on Github as you can see I've created my account only 8 days ago, so I don't know if it is the best practice but when I click "preview changes" it just works fine. Maybe this can solve the problem who knows... so please let me know... Also, I think that struggling with these problems its the best method to pragmatically understand how GitHub work and maybe I can learn something from it  ("you learn by your mistakes" or so they say...).


fixing images (added a single slash before "readme-only")

![Commit changes](/readme-only/image.PNG "description")
instead of 
![Commit changes](readme-only/image.PNG "description")

fixing links (remove the space between square and round brackets)

![Commit changes](readme-only/image.PNG "description")
instead of
![Commit changes] (readme-only/image.PNG "description")